kloudvin/api/requirements.txt
# Node.js functions don't use this, but Python functions do
azure-functions
markitdown[docx]
azure-storage-blob
